+++ docs/changelog.rst
@@ -4,6 +4,13 @@ Change log
 Unreleased_
 -----------
 
+Added
+~~~~~
+
+* Support setting a level on each ``Logger``.
+* Function to get the global log level.
+* Function to get the effective log level on a ``Logger``.
+
 Changed
 ~~~~~~~