+++ docs/changelog.rst
@@ -4,6 +4,11 @@ Change log
 Unreleased_
 -----------
 
+3.0.0_ |--| 2022-11-03
+----------------------
+
+Support setting a level on each ``Logger``.
+
 Added
 ~~~~~
 
@@ -200,7 +205,8 @@ Added
 
 .. |--| unicode:: U+2013 .. EN DASH
 
-.. _Unreleased: https://github.com/nomis/mcu-uuid-log/compare/2.3.0...HEAD
+.. _Unreleased: https://github.com/nomis/mcu-uuid-log/compare/3.0.0...HEAD
+.. _3.0.0: https://github.com/nomis/mcu-uuid-log/compare/2.3.0...3.0.0
 .. _2.3.0: https://github.com/nomis/mcu-uuid-log/compare/2.2.0...2.3.0
 .. _2.2.0: https://github.com/nomis/mcu-uuid-log/compare/2.1.4...2.2.0
 .. _2.1.4: https://github.com/nomis/mcu-uuid-log/compare/2.1.3...2.1.4